Move import completion to megaphone publisher
